VulkanRenderer - descriptors sets creation Renderer - now makes array with objects to render Shaders - compiling shaders on build

+++ ImGui/src/ImageBtnText.cpp
@@ -31,11 +31,11 @@ bool ImageButtonWithText(ImTextureID texId, const char* label, unsigned int* _si
     ImVec2 padding = (frame_padding >= 0) ? ImVec2((float)frame_padding, (float)frame_padding) : style.FramePadding;
 
 
-    int text_end = padding.x + textSize.x;
-    int addv = text_end > size.x + innerSpacing ? text_end - size.x - innerSpacing : 0;
+    int text_end = (int)(padding.x + textSize.x);
+    int addv = text_end > (int)size.x + (int)innerSpacing ? text_end - (int)size.x - (int)innerSpacing : 0;
 
     const ImVec2 totalSizeWithoutPadding(size.x + innerSpacing + addv, size.y + textSize.y);
-    *_size = totalSizeWithoutPadding.x;
+    *_size = (int)totalSizeWithoutPadding.x;
 
     const ImRect bb(window->DC.CursorPos, window->DC.CursorPos + totalSizeWithoutPadding + padding * 2);
     ImVec2 start(0, 0);
